sophia/.openclaw/extensions/compound-engineering/skills/gemini-imagegen/requirements.txt
google-genai>=1.0.0
Pillow>=10.0.0
